checkAvailableSpace method get a double not an it for fileSize

+++ src/main/java/dao/UserFileDao.java
@@ -21,5 +21,5 @@ public interface UserFileDao {
 
     int checkUserFile(UserFile userFile);
 
-    boolean checkAvailableSpace(int id, int fileSize);
+    boolean checkAvailableSpace(int id, double fileSize);
 }
